Use one socket fd for an instance
The UDPSocket.new for every stat is inefficient and makes it difficult
to debug with strace since every stat that is sent requires opening
files, dns lookups, etc.

This removes the file open/close and then gives an option of
`:lookup_always` to remove unneeded dns queries in the case of localhost
or a static A record.

I added a :resolve_always constructor option which does the following:
  - keeps backwards compatible API
  - forces sane defaults when not specified:
    - if localhost: default to only resolve the address on instantiation
    - if non-localhost: default to resolve every send_stats call
  - allows overriding the default behavior by simply setting the option

Documented punting thread-safety during initialization in the README, as
well as new :resolve_always.

Added a `connect` call to avoid resolving the host on every call to
`send`.

Bumps version to 1.1.0

Fixes #8